requirements.txt
aiohttp>=3.7.4.post0
asyncio>=3.4.3
beautifulsoup4>=4.9.3
PyQt5>=5.15.2
requests>=2.25.1
PyYAML>=6.0
